Add test for css3 transforms: rotate. Need further review.
The original ref file from JiKai uses two triangles(by some border
tricks) to make a rotated square. However, it's quite hard to determine
the position and size of the triangles because they are not integers. As
a result, the test and the ref cannot be rendered identically.

I wrote a new ref file using SVG and transform attribute on a rect
element. I'm not very sure if this can be seen as a different feature,
so I leave the test unapproved.