Make testharness tests work in Blink's test runner
Blink's test scripts only accept a testharness.js test as passing if
the only lines of text output from the test are output from
testharness.js itself. So this hides the visual part of the test and
only displays the testharness output to make these tests fully
compatible with running in Blink.