[css-fonts] Test font style matching clauses 4 a-c
This test exercises the font style matching clauses 4. a-c by iterating
over all possible combinations of stretch, style and weight. It declares
a @font-face rules and a style declaration for each combination, then
uses it in the HTML.

Each @font-face rule refers to a different ttf files with a special
property: it contains ligatures for the words stretch, style and weight
and replaces those words with the respective value.

This allows verification of whether the UA selected the right font face
rule and loaded the right corresponding ttf.

The test fonts were generated using custom scripts and the fonttools
library by subsetting, reducing and modifying the ligatures lookup in
the Linux Libertine Font, which is dual licensed under GPL3 and OFL.