Add a css-variables test that Firefox fails.
Firefox represents custom property values as strings internally,
and inserts `/**/` between adjacent tokens that would otherwise
be re-parsed as a single token.

It relies on the table [1] in the css-syntax spec,
but that table has not been updated when the definition
of identifier tokens was extended [2] to allow `--` prefixes.
This was just fixed [3].

[1] https://drafts.csswg.org/css-syntax/#serialization
[2] http://hg.csswg.org/drafts/rev/ab8b75f1fa88
[3] http://hg.csswg.org/drafts/rev/92d088e07ca2