Remove unbuilt CSS1 testsuite
Per the README, all the build system does is convert the XML source
files to XHTML 1.1 and adds link elements.

Given we're building into a single format and nobody is ever going to
use the source files for anything else, we can just get rid of the
unbuilt files and the build system entirely and allow people to edit
the built files.